+++ spec/fixtures/glimmer/dsl/swt/swt_dynamic_expression.rb
@@ -10,7 +10,7 @@ module Glimmer
         include ParentExpression
         
         def can_interpret?(parent, keyword, *args, &block)
-          return true unless ['html', 'body', 'input', 'color', 'font_size'].include?(keyword)
+          return true unless ['html', 'body', 'input', 'color', 'font_size', 'text'].include?(keyword)
         end
         
         def interpret(parent, keyword, *args, &block)